requirements.txt
streamlit~=1.43.2
requests~=2.32.3
python-dotenv~=1.1.0
mistralai~=1.6.0